refactor(unit_tests): remove duplicated code for DummyNode
Task:
https://trello.com/c/sFG8mzOe/4077-refactor-remove-duplicated-code-for-dummynode
